It's the final countdown
Come on, London

Come on, come on

We′re leaving together
But still it's farewell
And maybe we′ll come back
To earth, who can tell?

I guess there is no one to blame
We're leaving ground
Will things ever be the same again?

It's the final countdown
Come on, come on
The final countdown

Oh, we′re heading for Venus, Venus
Still we stand tall
′Cause maybe they've seen us
And welcome us all

So many light years to go
And things to be found (to be found)
I′m sure that we'll all miss her so, come on

(It′s the final countdown)
The final countdown
It's the final countdown (the final countdown)
Woah, yeah

It′s the final countdown
Woah, oh
It's the final countdown
Come on
The final countdown
It's the final countdown (the final countdown)
Woah

It′s the final countdown
We′re leaving together
The final countdown
We'll all miss her so
It′s the final countdown (the final countdown)
Woah, yeah

Woah, woah, woah

Alright